requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
mysql-connector-python
sqlalchemy
pinecone-client
langgraph
langchain
langchain-google-genai
google-generativeai
beautifulsoup4
python-dotenv
linkedin-api
requests
asyncmy
aiohttp
google-genai
python-dotenv